+++ EZVIDIA/EZVIDIA.vcxproj
@@ -143,9 +143,9 @@
     <ClCompile Include="src\networking\TcpServer.cpp" />
     <ClCompile Include="src\networking\TcpSocket.cpp" />
     <ClCompile Include="src\utils\StringUtils.cpp" />
-    <ClCompile Include="WindowsGui.cpp" />
-    <ClCompile Include="WindowsUtils.cpp" />
-    <ClCompile Include="WinMain.cpp" />
+    <ClCompile Include="src\gui\WindowsGui.cpp" />
+    <ClCompile Include="src\utils\WindowsUtils.cpp" />
+    <ClCompile Include="src\WinMain.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Image Include="logo.ico" />
@@ -170,10 +170,10 @@
     <ClInclude Include="src\drivers\NvapiPathInfoWrapper.hpp" />
     <ClInclude Include="src\drivers\DisplayDriver.hpp" />
     <ClInclude Include="src\networking\NetworkException.hpp" />
-    <ClInclude Include="resource.h" />
-    <ClInclude Include="targetver.h" />
-    <ClInclude Include="WindowsGui.hpp" />
-    <ClInclude Include="WindowsUtils.hpp" />
+    <ClInclude Include="src\resource.h" />
+    <ClInclude Include="src\targetver.h" />
+    <ClInclude Include="src\gui\WindowsGui.hpp" />
+    <ClInclude Include="src\utils\WindowsUtils.hpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\README.md" />
