--- EZVIDIA/resource.h
@@ -1,33 +1,0 @@
-//{{NO_DEPENDENCIES}}
-// Microsoft Visual C++ generated include file.
-// Used by ezvidia.rc
-//
-#define IDC_MYICON                      2
-#define IDD_EZVIDIASERVERDEV_DIALOG     102
-#define IDS_APP_TITLE                   103
-#define IDM_EXIT                        105
-#define IDI_EZVIDIASERVERDEV            107
-#define IDI_SMALL                       108
-#define IDC_EZVIDIASERVERDEV            109
-#define IDR_MAINFRAME                   128
-#define IDD_DIALOG1                     129
-#define IDM_CONFIGNUM                   150
-#define IDC_EDIT1                       1000
-#define IDM_DELCONFNUM                  1500
-#define IDM_SAVECONF                    2000
-#define IDM_GENERATEBAT                 2500
-#define IDM_NETWORK_ON					2501
-#define IDM_NETWORK_OFF					2502
-#define IDC_STATIC                      -1
-
-// Next default values for new objects
-//
-#ifdef APSTUDIO_INVOKED
-#ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NO_MFC                     1
-#define _APS_NEXT_RESOURCE_VALUE        131
-#define _APS_NEXT_COMMAND_VALUE         32771
-#define _APS_NEXT_CONTROL_VALUE         1002
-#define _APS_NEXT_SYMED_VALUE           110
-#endif
-#endif
